requirements.txt
Flask
SQLAlchemy
psycopg2-binary
python-dotenv
Pillow
flasgger  # <--- Добавьте эту строку